app/requirements.txt
python-telegram-bot>=21.0
httpx>=0.27.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
